% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combin8R-package.R
\docType{package}
\name{Combin8R-package}
\alias{Combin8R}
\alias{Combin8R-package}
\title{Combin8R: Parser Combinators for R}
\description{
A parser combinator package to assist the construction of recursive descent parsers. A parser is a function that accepts a string as input, and upon success returns a list containing any unconsumed input and the result of the parse. A parser combinator is a higher order function that combines one or more simple parsers to form a more complex parser.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SWotherspoon/Combin8R}
  \item Report bugs at \url{https://github.com/SWotherspoon/Combin8R/issues}
}

}
\author{
\strong{Maintainer}: Simon Wotherspoon \email{Simon.Wotherspoon@aad.gov.au} (\href{https://orcid.org/0000-0002-6947-4445}{ORCID})

Authors:
\itemize{
  \item Michael Sumner \email{mdsumner@gmail.com} (\href{https://orcid.org/0000-0002-2471-7511}{ORCID})
}

}
\keyword{internal}
