% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combin8R.R
\name{formatAST}
\alias{formatAST}
\title{Format an Abstract Syntax Tree}
\usage{
formatAST(x, indent = 0)
}
\arguments{
\item{x}{The result of a parser.}

\item{indent}{The initial indentation level.}
}
\value{
A string representation of the abstract syntax tree
}
\description{
Format an abstract syntax tree as a string
}

