% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combin8R.R
\name{pAlt}
\alias{pAlt}
\title{"Alt" Parser Combinator}
\usage{
pAlt(tag, ...)
}
\arguments{
\item{tag}{a class name or a function of a single argument that
constructs an object representing the result of the parse.}

\item{...}{Alternative simpler parsers.}
}
\value{
Returns a parser.
}
\description{
Creates a parser that accepts input that is accepted by any one of
a number of simpler parsers.
}
\details{
The \code{tag} argument should either be a string naming the class
of the result, or a function that constructs a object representing
the result of the parse given the result of the first alternative
parser satisfied by the input.
}
