% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combin8R.R
\name{pLiteral}
\alias{pLiteral}
\title{Create a Primitive Parser from a Regex}
\usage{
pLiteral(tag, string = tag)
}
\arguments{
\item{tag}{a class name or a function of a single argument that
constructs an object representing the result of the parse.}

\item{string}{The string literal to match.}
}
\value{
Returns a parser.
}
\description{
Create a parser to parse text matching a literal string.
}
\details{
This function creates a parser that accepts input that matches a
given string.

The \code{tag} argument should either be a string naming the class
of the result, or a function that constructs a object representing
the result of the parse, given the matched literal string.
}
