% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combin8R.R
\name{pMany}
\alias{pMany}
\title{"Many" Parser Combinator}
\usage{
pMany(tag, p)
}
\arguments{
\item{tag}{a class name or a function of a single argument that
constructs an object representing the result of the parse.}

\item{p}{A parser to match zero or more times.}
}
\value{
Return a parser.
}
\description{
Create a parser that succeeds on input for which a simpler parser
succeeds zero or more times in succession.
}
\details{
The \code{tag} argument should either be a string naming the class
of the result, or a function that constructs a object representing
the result of the parse given the list of results of the
constituent parser.
}

