% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combin8R.R
\name{pSeq}
\alias{pSeq}
\title{"Seq" Parser Combinator}
\usage{
pSeq(tag, ...)
}
\arguments{
\item{tag}{a class name or a function of a single argument that
constructs an object representing the result of the parse.}

\item{...}{A sequence of parsers to matched in sequence.}
}
\value{
Return a parser.
}
\description{
Create a parser that succeeds on input accepted by a sequence of
simpler parsers applied in succession.
}
\details{
The \code{tag} argument should either be a string naming the class
of the result, or a function that constructs a object representing
the result of the parse given the list of results of the
constituent parsers.
}

