% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{Combin8R-package}
\alias{Combin8R-package}
\title{Parser Combinators}
\description{
A parser combinator package to assist the construction of recursive
descent parsers. A parser is a function that accepts a string as
input, and upon success returns a list containing any unconsumed
input and the result of the parse.  A parser combinator is a higher
order function that combines one or more simple parses to form a
more complex parser.
}
\author{
Simon Wotherspoon, Micheal Sumner
}

